commit before switch
